Classify DistinctTile as composite tile
